+++ src/Components/Widget/BooleanWidget/index.tsx
@@ -3,7 +3,7 @@ import { Switch, Platform } from 'react-native';
 
 
 import { Layout } from '@Components/Layout';
-import { BooleanWidgetData } from '@Types/index';
+import { BooleanWidgetData, GPS_DTO } from '@Types/index';
 import { translations } from '@Translations/index';
 import ConfigService from '@Services/ConfigService';
 
@@ -20,30 +20,45 @@ export default function BooleanWidget(props: {
   const stringResources = useMemo(() => translations.Widgets.BooleanWidget[language], []);
 
   const [widgetData, setWidgetData] = useState<BooleanWidgetData>(props.widgetData);
+  const [showGPS, setShowGPS] = useState<boolean>(widgetData.gps !== undefined);
   const [showModal, setShowModal] = useState<boolean>(false);
   const [isDataWrong, setIsDataWrong] = useState<boolean>(false);
-
-  function onSwitchChange() {
-    setWidgetData(prev => {
-      const newWidgetData = {
-        ...prev,
-        value: !prev.value,
-        rules: {...prev.rules},
-      };
-      props.onConfirm(newWidgetData);
-      return newWidgetData;
-    });
-  }
 
   function onConfirm(widgetData: BooleanWidgetData) {
 
     setShowModal(false);
+
+    widgetData.gps !== undefined
+    ? setShowGPS(true)
+    : setShowGPS(false);
 
     if (showModal) {
       setIsDataWrong(false);
       setWidgetData(widgetData);
       props.onConfirm(widgetData);
     }
+  }
+
+  function onSwitchChange(boolean: boolean) {
+    const newWidgetData = { ...widgetData, value: boolean };
+    setWidgetData(newWidgetData);
+    props.onConfirm(newWidgetData);
+  }
+
+  function onSaveGPS(newGPSData: GPS_DTO) {
+    const newWidgetData = { ...widgetData, gps: newGPSData };
+    setWidgetData(newWidgetData);
+    props.onConfirm(newWidgetData);
+  }
+
+  function onDeleteGPS() {
+    if (widgetData.gps !== undefined) {
+      const newWidgetData = { ...widgetData };
+      delete newWidgetData.gps;
+      setWidgetData(newWidgetData);
+      props.onConfirm(newWidgetData);
+    }
+    setShowGPS(false);
   }
 
   return (<>
@@ -57,7 +72,9 @@ export default function BooleanWidget(props: {
       iconButtons_Top={
         <IconButtons_Top
           widgetData={widgetData}
+          showGPS={showGPS}
           onPencilPress={() => setShowModal(true)}
+          onGPSPress={() => setShowGPS(true)}
         />
       }
 
@@ -95,9 +112,21 @@ export default function BooleanWidget(props: {
           trackColor={{ false: theme.wrong, true: theme.confirm }}
           ios_backgroundColor={theme.wrong}
           value={widgetData.value}
-          onValueChange={() => onSwitchChange()}
+          onValueChange={(boolean) => onSwitchChange(boolean)}
         />
       </Layout.View>
+      {showGPS && (
+        <Layout.Input.GPS
+          label="GPS"
+          initialGPSData={widgetData.gps ?? {}}
+          backgroundColor={theme.tertiary}
+          color={theme.onTertiary}
+          color_placeholder={theme.onBackground_Placeholder}
+          onPress_Delete={() => onDeleteGPS()}
+          onPress_Save={(newGPSData) => onSaveGPS(newGPSData)}
+          style={{ marginBottom: 10 }}
+        />
+      )}
 
     </WidgetComponent.Root>
   </>);
@@ -105,17 +134,32 @@ export default function BooleanWidget(props: {
 
 function IconButtons_Top(props: {
   widgetData: BooleanWidgetData
+  showGPS: boolean
   onPencilPress: () => void
+  onGPSPress: () => void
 }) {
 
   const { theme } = useMemo(() => ConfigService.config, []);
 
   return (<>
+    {!props.showGPS && (
+      <Layout.Button.Icon
+        iconName="location"
+        color_background={theme.secondary}
+        color={theme.onSecondary}
+        onPress={() => props.onGPSPress()}
+        style={{
+          height: 40,
+          paddingHorizontal: 10,
+          paddingVertical: 5,
+        }}
+      />
+    )}
     {(props.widgetData.rules.allowLabelChange) && (
       <Layout.Button.Icon
         iconName="pencil-sharp"
         color={theme.onSecondary}
-        onPress={props.onPencilPress}
+        onPress={() => props.onPencilPress()}
         style={{
           paddingHorizontal: 10,
           paddingVertical: 5,
@@ -138,24 +182,34 @@ function Modal(props: {
 
   const [label, setLabel] = useState<string>(props.widgetData.name);
   const [value, setValue] = useState<boolean>(props.widgetData.value);
+  const [gpsData, setGPSData] = useState<GPS_DTO | undefined>(props.widgetData.gps);
 
   const { rules } = props.widgetData;
+
+  function onConfirm() {
+
+    const newWidgetData: BooleanWidgetData = {
+      id_widget: props.widgetData.id_widget,
+      name: label,
+      type: 'boolean',
+      value: value,
+      rules: { ...props.widgetData.rules },
+    };
+
+    if (gpsData !== undefined) {
+      newWidgetData.gps = gpsData;
+    }
+
+    props.onConfirm(newWidgetData);
+  }
 
   return (
     <WidgetComponent.Modal
       title={label}
       widgetData={props.widgetData}
-      onConfirm={() => {
-        props.onConfirm({
-          id_widget: props.widgetData.id_widget,
-          name: label,
-          type: 'boolean',
-          value: value,
-          rules: { ...props.widgetData.rules },
-        });
-      }}
-      onDelete={props.onDelete}
-      onRequestClose={props.onRequestClose}
+      onConfirm={() => onConfirm()}
+      onDelete={() => props.onDelete()}
+      onRequestClose={() => props.onRequestClose()}
     >
       <Layout.View
         style={{
@@ -180,6 +234,36 @@ function Modal(props: {
           locked={!rules.allowValueChange}
           onSwitchChange={(boolean) => setValue(boolean)}
         />
+        {gpsData === undefined ? (
+          <Layout.View
+            style={{
+              alignItems: 'flex-end',
+              marginHorizontal: 10,
+            }}
+          >
+            <Layout.Button.Icon
+              iconName="location"
+              color_background={theme.secondary}
+              color={theme.onSecondary}
+              onPress={() => setGPSData({})}
+              style={{
+                borderRadius: 10,
+                height: 40,
+                width: 80,
+              }}
+            />
+          </Layout.View>
+        ) : (
+          <Layout.Input.GPS
+            label="GPS"
+            initialGPSData={gpsData}
+            backgroundColor={theme.background}
+            color={theme.onBackground}
+            color_placeholder={theme.onBackground_Placeholder}
+            onPress_Delete={() => setGPSData(undefined)}
+            onPress_Save={(newGPSData) => setGPSData(newGPSData)}
+          />
+        )}
       </Layout.View>
     </WidgetComponent.Modal>
   );