+++ src/Components/Widget/TextWidget/index.tsx
@@ -2,7 +2,7 @@ import React, { useMemo, useState } from 'react';
 import { WidgetComponent } from '@WidgetComponents/index';
 
 import { Layout } from '@Components/Layout';
-import { TextWidgetData, WidgetData } from '@Types/index';
+import { GPS_DTO, TextWidgetData, WidgetData } from '@Types/index';
 import { translations } from '@Translations/index';
 import ConfigService from '@Services/ConfigService';
 
@@ -19,13 +19,12 @@ export default function TextWidget(props: {
   const { theme, language } = useMemo(() => ConfigService.config, []);
   const stringResources = useMemo(() => translations.Widgets.TextWidget[language], []);
 
-  const [showModal, setShowModal] = useState<boolean>(false);
   const [widgetData, setWidgetData] = useState<TextWidgetData>(UtilService.deepCloning(props.widgetData));
+  const [showGPS, setShowGPS] = useState<boolean>(widgetData.gps !== undefined);
+  const [showModal, setShowModal] = useState<boolean>(false);
   const [isDataWrong, setIsDataWrong] = useState<boolean>(false);
 
-  function onConfirm(widgetData: TextWidgetData) {
-
-    setShowModal(false);
+  function validateConfirm(widgetData: TextWidgetData) {
 
     if (WidgetRules.noSpaces(widgetData)) {
       alert(stringResources['Value cannot have empty spaces.']);
@@ -40,12 +39,19 @@ export default function TextWidget(props: {
     }
 
     setIsDataWrong(false);
-    setWidgetData(widgetData);
     props.onConfirm(widgetData);
   }
 
   function onConfirm_Modal(widgetData: TextWidgetData) {
-    onConfirm(widgetData);
+
+    setShowModal(false);
+
+    widgetData.gps !== undefined
+    ? setShowGPS(true)
+    : setShowGPS(false);
+
+    setWidgetData(widgetData);
+    validateConfirm(widgetData);
   }
 
   function onTextChange(text: string) {
@@ -57,7 +63,23 @@ export default function TextWidget(props: {
       },
     };
     setWidgetData(newData);
-    onConfirm(newData);
+    validateConfirm(newData);
+  }
+
+  function onSaveGPS(newGPSData: GPS_DTO) {
+    const newWidgetData = { ...widgetData, gps: newGPSData };
+    setWidgetData(newWidgetData);
+    props.onConfirm(newWidgetData);
+  }
+
+  function onDeleteGPS() {
+    if (widgetData.gps !== undefined) {
+      const newWidgetData = { ...widgetData };
+      delete newWidgetData.gps;
+      setWidgetData(newWidgetData);
+      props.onConfirm(newWidgetData);
+    }
+    setShowGPS(false);
   }
 
   return (
@@ -71,7 +93,9 @@ export default function TextWidget(props: {
       iconButtons_Top={
         <IconButtons_Top
           widgetData={widgetData}
+          showGPS={showGPS}
           onPencilPress={() => setShowModal(true)}
+          onGPSPress={() => setShowGPS(true)}
         />
       }
 
@@ -102,6 +126,17 @@ export default function TextWidget(props: {
           locked={false}
           onChangeText={(text) => onTextChange(text)}
         />
+        {showGPS && (
+          <Layout.Input.GPS
+            label="GPS"
+            initialGPSData={widgetData.gps ?? {}}
+            backgroundColor={theme.tertiary}
+            color={theme.onTertiary}
+            color_placeholder={theme.onBackground_Placeholder}
+            onPress_Delete={() => onDeleteGPS()}
+            onPress_Save={(newGPSData) => onSaveGPS(newGPSData)}
+          />
+        )}
       </Layout.View>
 
     </WidgetComponent.Root>
@@ -110,12 +145,27 @@ export default function TextWidget(props: {
 
 function IconButtons_Top(props: {
   widgetData: TextWidgetData
+  showGPS: boolean
   onPencilPress: () => void
+  onGPSPress: () => void
 }) {
 
   const { theme } = useMemo(() => ConfigService.config, []);
 
   return (<>
+    {!props.showGPS && (
+      <Layout.Button.Icon
+        iconName="location"
+        color_background={theme.secondary}
+        color={theme.onSecondary}
+        onPress={() => props.onGPSPress()}
+        style={{
+          height: 40,
+          paddingHorizontal: 10,
+          paddingVertical: 5,
+        }}
+      />
+    )}
     {(props.widgetData.rules.allowLabelChange || props.widgetData.rules.allowValueChange) && (
       <Layout.Button.Icon
         iconName="pencil-sharp"
@@ -143,22 +193,32 @@ function Modal(props: {
 
   const [label, setLabel] = useState<string>(props.widgetData.name);
   const [value, setValue] = useState<string>(props.widgetData.value);
+  const [gpsData, setGPSData] = useState<GPS_DTO | undefined>(props.widgetData.gps);
 
   const { rules } = props.widgetData;
+
+  function onConfirm() {
+
+    const newWidgetData: TextWidgetData = {
+      id_widget: props.widgetData.id_widget,
+      name: label,
+      type: 'text',
+      value: value,
+      rules: { ...props.widgetData.rules },
+    };
+
+    if (gpsData !== undefined) {
+      newWidgetData.gps = gpsData;
+    }
+
+    props.onConfirm(newWidgetData);
+  }
 
   return (
     <WidgetComponent.Modal
       title={label}
       widgetData={props.widgetData}
-      onConfirm={() => {
-        props.onConfirm({
-          id_widget: props.widgetData.id_widget,
-          name: label,
-          type: 'text',
-          value: value,
-          rules: { ...rules },
-        });
-      }}
+      onConfirm={() => onConfirm()}
       onDelete={props.onDelete}
       onRequestClose={props.onRequestClose}
     >
@@ -187,6 +247,37 @@ function Modal(props: {
           locked={!rules.allowValueChange}
           onChangeText={(text) => setValue(text)}
         />
+        {gpsData === undefined ? (
+          <Layout.View
+            style={{
+              alignItems: 'flex-end',
+              marginHorizontal: 10,
+            }}
+          >
+            <Layout.Button.Icon
+              iconName="location"
+              color_background={theme.secondary}
+              color={theme.onSecondary}
+              onPress={() => setGPSData({})}
+              style={{
+                borderRadius: 10,
+                height: 40,
+                width: 80,
+              }}
+            />
+          </Layout.View>
+        ) : (
+          <Layout.Input.GPS
+            label="GPS"
+            initialGPSData={gpsData}
+            backgroundColor={theme.background}
+            color={theme.onBackground}
+            color_placeholder={theme.onBackground_Placeholder}
+            onPress_Delete={() => setGPSData(undefined)}
+            onPress_Save={(newGPSData) => setGPSData(newGPSData)}
+            style={{ borderRadius: 10 }}
+          />
+        )}
       </Layout.View>
     </WidgetComponent.Modal>
   );