+++ src/app/ProjectCreationScreen/LocalComponents/API_Inputs_ProjectSettings.ts
@@ -1,7 +1,14 @@
+import { GPS_DTO } from '@Types/index';
 import API_TemporaryProject from './API_TemporaryProject';
 
 export default class API_Inputs_ProjectSettings {
 
-  static setProjectID(value: string) { API_TemporaryProject.project.projectSettings.id_project = value; }
-  static setProjectName(value: string) { API_TemporaryProject.project.projectSettings.name = value; }
+  static setID(value: string) { API_TemporaryProject.project.projectSettings.id_project = value; }
+  static setName(value: string) { API_TemporaryProject.project.projectSettings.name = value; }
+  static setGPS(gpsData: GPS_DTO) { API_TemporaryProject.project.projectSettings.gps = gpsData; }
+  static deleteGPS() {
+    if (API_TemporaryProject.project.projectSettings.gps === undefined) {
+      delete API_TemporaryProject.project.projectSettings.gps;
+    }
+  }
 }