+++ src/app/ProjectCreationScreen/LocalComponents/Inputs_ProjectSettings.tsx
@@ -7,6 +7,7 @@ import ConfigService from '@Services/ConfigService';
 
 import API_TemporaryProject from './API_TemporaryProject';
 import API_Inputs_ProjectSettings from './API_Inputs_ProjectSettings';
+import { GPS_DTO } from '@Types/index';
 
 export default function Inputs_ProjectSettings() {
 
@@ -15,38 +16,65 @@ export default function Inputs_ProjectSettings() {
 
   const [id, setId] = useState<string>(API_TemporaryProject.project.projectSettings.id_project);
   const [name, setName] = useState<string>(API_TemporaryProject.project.projectSettings.name);
+  const [gpsData, setGPSData] = useState<GPS_DTO | undefined>(API_TemporaryProject.project.projectSettings.gps);
+  const [showGPS, setShowGPS] = useState<boolean>(false);
 
   function onIDChange(newID: string) {
     const normalizedText = newID.replace(UtilService.idRegex, '');
-    API_Inputs_ProjectSettings.setProjectID(normalizedText);
+    API_Inputs_ProjectSettings.setID(normalizedText);
     setId(normalizedText);
   }
 
   function onNameChange(newName: string) {
-    API_Inputs_ProjectSettings.setProjectName(newName);
+    API_Inputs_ProjectSettings.setName(newName);
     setName(newName);
+  }
+
+  function onSaveGPS(gpsData: GPS_DTO) {
+    API_Inputs_ProjectSettings.setGPS(UtilService.deepCloning(gpsData));
+    setGPSData(UtilService.deepCloning(gpsData));
+  }
+
+  function onDeleteGPS() {
+    API_Inputs_ProjectSettings.deleteGPS();
+    setGPSData(undefined);
+    setShowGPS(false);
   }
 
   return (
     <Layout.View>
       <Layout.View
         style={{
+          flexDirection: 'row',
+          justifyContent: 'space-between',
           backgroundColor: theme.secondary,
-          height: 40,
-          padding: 5,
           borderTopLeftRadius: 10,
           borderTopRightRadius: 10,
         }}
       >
         <Layout.Text.P
           style={{
-            paddingVertical: 5,
-            paddingHorizontal: 10,
+            paddingVertical: 10,
+            paddingHorizontal: 15,
             color: theme.onSecondary,
           }}
         >
           {stringResources['Project info']}
         </Layout.Text.P>
+        {!showGPS && (
+          <Layout.Button.Icon
+            iconName="location"
+            color_background={theme.secondary}
+            color={theme.onSecondary}
+            onPress={() => setShowGPS(true)}
+            style={{
+              height: 40,
+              borderTopRightRadius: 10,
+              paddingHorizontal: 10,
+              paddingVertical: 5,
+            }}
+          />
+        )}
       </Layout.View>
       <Layout.View
         style={{
@@ -78,6 +106,17 @@ export default function Inputs_ProjectSettings() {
           locked={false}
           onChangeText={(text) => onNameChange(text)}
         />
+        {showGPS && (
+          <Layout.Input.GPS
+            label="GPS"
+            initialGPSData={gpsData ?? {}}
+            backgroundColor={theme.tertiary}
+            color={theme.onTertiary}
+            color_placeholder={theme.onBackground_Placeholder}
+            onPress_Delete={() => onDeleteGPS()}
+            onPress_Save={(newGPSData) => onSaveGPS(newGPSData)}
+          />
+        )}
       </Layout.View>
     </Layout.View>
   );