+++ src/app/[id_project]/ProjectSettingsScreen/LocalComponents/Inputs_ProjectSettings.tsx
@@ -4,7 +4,7 @@ import { useLocalSearchParams } from 'expo-router';
 import { Layout } from '@Components/Layout';
 import { useTimeout } from '@Hooks/index';
 import { translations } from '@Translations/index';
-import { ProjectSettings } from '@Types/index';
+import { GPS_DTO, ProjectSettings } from '@Types/index';
 import ConfigService from '@Services/ConfigService';
 import ProjectService from '@Services/ProjectService';
 import CacheService from '@Services/CacheService';
@@ -18,6 +18,7 @@ export default function Inputs_ProjectSettings() {
   const stringResources = useMemo(() => translations.Screens.ProjectSettingsScreen[language], []);
 
   const [projectSettings, setProjectSettings] = useState<ProjectSettings>(UtilService.deepCloning(CacheService.getProjectFromCache(id_project)));
+  const [showGPS, setShowGPS] = useState<boolean>(projectSettings.gps !== undefined);
   const [saved, setSaved] = useState<boolean>(true);
   const { rules } = projectSettings;
 
@@ -27,13 +28,24 @@ export default function Inputs_ProjectSettings() {
 
   function onNameChange(newName: string) {
     if (rules.allowNameChange) {
-      setProjectSettings(prev => {
-        const newData = { ...prev };
-        newData.name = newName;
-        return newData;
-      });
+      setProjectSettings(prev => ({ ...prev, name: newName }));
       setSaved(false);
     }
+  }
+
+  function onSaveGPS(newGPSData: GPS_DTO) {
+    setProjectSettings(prev => ({ ...prev, gps: newGPSData }));
+    setSaved(false);
+  }
+
+  function onDeleteGPS() {
+    if (projectSettings.gps !== undefined) {
+      const newProjectSettings = { ...projectSettings };
+      delete newProjectSettings.gps;
+      setProjectSettings(newProjectSettings);
+      setSaved(false);
+    }
+    setShowGPS(false);
   }
 
   return (
@@ -41,28 +53,48 @@ export default function Inputs_ProjectSettings() {
       <Layout.View
         style={{
           flexDirection: 'row',
-          alignItems: 'center',
+          justifyContent: 'space-between',
           backgroundColor: theme.secondary,
-          height: 40,
-          padding: 5,
-          paddingHorizontal: 10,
           borderTopLeftRadius: 10,
           borderTopRightRadius: 10,
         }}
       >
-        <Layout.StatusFeedback
-          done={saved}
-          error={false}
-        />
-        <Layout.Text.P
+        <Layout.View
           style={{
+            flexDirection: 'row',
+            alignItems: 'center',
             paddingVertical: 5,
             paddingHorizontal: 10,
-            color: theme.onSecondary,
           }}
         >
-          {stringResources['Project info']}
-        </Layout.Text.P>
+          <Layout.StatusFeedback
+            done={saved}
+            error={false}
+          />
+          <Layout.Text.P
+            style={{
+              paddingVertical: 5,
+              paddingHorizontal: 10,
+              color: theme.onSecondary,
+            }}
+          >
+            {stringResources['Project info']}
+          </Layout.Text.P>
+        </Layout.View>
+        {!showGPS && (
+          <Layout.Button.Icon
+            iconName="location"
+            color_background={theme.secondary}
+            color={theme.onSecondary}
+            onPress={() => setShowGPS(true)}
+            style={{
+              height: 40,
+              borderTopRightRadius: 10,
+              paddingHorizontal: 10,
+              paddingVertical: 5,
+            }}
+          />
+        )}
       </Layout.View>
       <Layout.View
         style={{
@@ -94,6 +126,17 @@ export default function Inputs_ProjectSettings() {
           locked={!rules.allowNameChange}
           onChangeText={(text) => onNameChange(text)}
         />
+        {showGPS && (
+          <Layout.Input.GPS
+            label="GPS"
+            initialGPSData={projectSettings.gps ?? {}}
+            backgroundColor={theme.tertiary}
+            color={theme.onTertiary}
+            color_placeholder={theme.onBackground_Placeholder}
+            onPress_Delete={() => onDeleteGPS()}
+            onPress_Save={(newGPSData) => onSaveGPS(newGPSData)}
+          />
+        )}
       </Layout.View>
     </Layout.View>
   );