+++ src/app/[id_project]/SampleCreationScreen/LocalComponents/API_Inputs_SampleSettings.ts
@@ -1,20 +1,18 @@
-import { SampleSettings } from '@Types/index';
+import { GPS_DTO, SampleSettings } from '@Types/index';
 import ProjectService from '@Services/ProjectService';
 
 export default class API_Inputs_SampleSettings {
 
   static temporarySettings: SampleSettings = ProjectService.getDefaultSampleSettings();
 
-  static reset() {
-    this.temporarySettings = ProjectService.getDefaultSampleSettings();
-  }
-
-  static setSampleID(value: string) {
-    this.temporarySettings.id_sample = value;
-  }
-
-  static setSampleName(value: string) {
-    this.temporarySettings.name = value;
+  static reset() { this.temporarySettings = ProjectService.getDefaultSampleSettings(); }
+  static setSampleID(value: string) { this.temporarySettings.id_sample = value; }
+  static setSampleName(value: string) { this.temporarySettings.name = value; }
+  static setGPS(gpsData: GPS_DTO) { this.temporarySettings.gps = gpsData; }
+  static deleteGPS() {
+    if (this.temporarySettings.gps === undefined) {
+      delete this.temporarySettings.gps;
+    }
   }
 }