+++ docs/ReleaseNotes.md
@@ -1,4 +1,10 @@
-# 0.1.1.10 GPSInput Translations
+# 0.1.1.11 Project, Sample, Widget GPSdata.
 
-- Removed delete alert message, so parent components can implement alert messages on their own.
-- Added Translations for GPSInput on Translations index
+- Added gps to projectSettings inputs on Project Creation Screen
+- Added gps to projectSettings inputs on Project Settings Screen
+- Added gps to projectSettings inputs on Sample Creation Screen
+- Added gps to projectSettings inputs on Sample Settings Screen
+- GPSWidget created.
+
+Fix:
+- Widget Modal: bottom padding inside Modal children container, causing sum with scrollview default bottom padding, causing a total of 300 bottom padding.